bit-rot: simplify and cleanup time management
Drop trivial 'br_fsscan_calculate_delta()', consistently use
'unsigned long' and 'time_t' to represent time intervals here
and there, drop unused 'tv' from 'struct br_child', adjust
related code.

Updates: #1000

+++ xlators/features/bit-rot/src/stub/bit-rot-common.h
@@ -103,8 +103,8 @@ typedef struct br_isignature_out {
 
     unsigned long version; /* current signed version    */
 
-    uint32_t time[2]; /* time when the object
-                         got dirtied               */
+    unsigned long time[2]; /* time when the object
+                              got dirtied               */
 
     int8_t signaturetype; /* hash type                 */
     size_t signaturelen;  /* signature length          */
@@ -112,7 +112,7 @@ typedef struct br_isignature_out {
 } br_isignature_out_t;
 
 typedef struct br_stub_init {
-    uint32_t timebuf[2];
+    unsigned long timebuf[2];
     char export[PATH_MAX];
 } br_stub_init_t;
 
@@ -139,7 +139,7 @@ br_is_signature_type_valid(int8_t signaturetype)
 }
 
 static inline void
-br_set_default_ongoingversion(br_version_t *buf, uint32_t *tv)
+br_set_default_ongoingversion(br_version_t *buf, unsigned long *tv)
 {
     buf->ongoingversion = BITROT_DEFAULT_CURRENT_VERSION;
     buf->timebuf[0] = tv[0];
@@ -156,7 +156,8 @@ br_set_default_signature(br_signature_t *buf, size_t *size)
 }
 
 static inline void
-br_set_ongoingversion(br_version_t *buf, unsigned long version, uint32_t *tv)
+br_set_ongoingversion(br_version_t *buf, unsigned long version,
+                      unsigned long *tv)
 {
     buf->ongoingversion = version;
     buf->timebuf[0] = tv[0];
